Neobotix version of kinetic-devel branch of their fork of ar_track_alvar
THIS IS THE NEOBOTIX VERSION OF THE kinetic-devel BRANCH OF ar_rack_alvar.
THEY HAVE MADE CHANGES IN THIS BRANCH IN ORDER FOR neo-docking TO WORK PROPERLY! 

THIS IS THE NEOBOTIX VERSION OF THE kinetic-devel BRANCH OF ar_rack_alvar.
THEY HAVE MADE CHANGES IN THIS BRANCH IN ORDER FOR neo-docking TO WORK PROPERLY!